requirements.txt
pandas
matplotlib
jupyterlab
scikit-learn
librosa 
soundfile
numpy
praat-parselmouth